NFC_Suite/requirements.txt
flask
flask-socketio
paho-mqtt
torch
transformers
pycryptodome
nfcpy
